did cart.js minimal move all to ShoppingCart.mjs
